requirements.txt
# Cubist - 3x3 Rubik's Cube Solver & Tutor
# Python 3.11+ required

# GUI Framework
PySide6>=6.6.0

# 3D Graphics
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7

# Fast Solver
kociemba>=1.2.1

# Math utilities
numpy>=1.24.0

# PDF Export
reportlab>=4.0.0

# Development & QA
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest>=7.4.0
pytest-qt>=4.2.0

# Packaging
PyInstaller>=5.13.0
